Add NextUp live activity
